requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
html5lib==1.1
idna==3.7
outcome==1.3.0.post0
pillow==10.4.0
PySocks==1.7.1
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
